ADR 027 Step 5: retire legacy storage chain, port cloud sync to Yjs
Delete the USE_YJS_STORAGE flag and legacy local chain (useSyncedStorage,
usePersistedStorage, useYjsToLegacyMirror, StorageFlagReloadBanner, the
bwp-storage-flag channel, the allotment-unified-data auto-save write path,
and the PR #369 same-tab broadcast apparatus). Collapse the eight domain
hooks to their mutate() branch only. Port the Supabase cloud-sync layer
(fetch/push/LWW/conflict + 30s debounce + unload flush) into useCloudSync,
consuming the Yjs snapshot directly. Rename src/lib/yjs-spike -> src/lib/yjs,
consolidate the write-site helpers, and rework the path-parity test into a
Yjs-path regression test.
