+++ docs/plans/current-plan.md
@@ -604,7 +604,17 @@ same tip deduped across areas), and the new advice-below-action ordering.
 Regression test added for the sort; type-check, lint, and the full unit suite
 pass.
 
-### Up Next: Phase 1 soak then Step 5 cleanup
+### ADR 027 Yjs Step 5 — Cleanup shipped
+
+The soak window passed with no data anomalies, so Step 5 deleted the legacy storage chain and its rollback scaffolding. `useAllotmentData` now composes `useYjsDoc` (canonical local engine, IndexedDB via `y-indexeddb`) with a new `useCloudSync` hook unconditionally — the `USE_YJS_STORAGE` flag is gone. The eight domain hooks (`useAllotmentAreas`, `useAllotmentPlantings`, `useAllotmentVarieties`, `useAllotmentCustomTasks`, `useAllotmentMaintenance`, `useAllotmentNotes`, `useAllotmentCareLogs`, `useCompost`) keep only their `mutate()` branch; the `setData` prop and legacy branches are removed.
+
+Deleted: `useSyncedStorage`, `usePersistedStorage`, `useYjsToLegacyMirror`, `StorageFlagReloadBanner`, the `bwp-storage-flag` BroadcastChannel, the legacy `allotment-unified-data` debounced-save write path, and the same-tab broadcast apparatus from PR #369 (`bonnie:storage-update` CustomEvent, `instanceId`/`sameTabSeq` bookkeeping, `recordSavedState`/`recordAdoptedState`, `recentSavesRef` echo dedup). The `allotment-unified-data` key is still *read* on first run and *written* by import/restore/fresh-init as the hydrate seed; only the mirror/auto-save write path is gone. Cross-tab sync now rides `y-indexeddb`'s IndexedDB broadcast.
+
+**Cloud sync preserved.** `useCloudSync` (`src/hooks/useCloudSync.ts`) ports the Supabase fetch/push/LWW-guard/conflict-dialog layer, the 30s push debounce, and the unload flush off `usePersistedStorage` to consume the Yjs snapshot directly. Push fires on Yjs snapshot-reference changes (the old chain used `saveStatus === 'saved'`); `'cloud'` conflict resolution adopts the remote via `useYjsDoc.replaceFromJson`, `'local'` pushes the current snapshot. A `normalizedContentSnapshot` round-trips the adopted remote through hydrate/serialize so the republished snapshot isn't pushed straight back. `serializeToJson`/`decodeDocState` stay forever (rollback + GDPR export + debug).
+
+Follow-ups folded in: `src/lib/yjs-spike/` → `src/lib/yjs/` (imports, CLAUDE.md, ADR 027 updated); `yjs-helpers.ts` (`withoutUndefined`/`assignDefined`) consolidated into `allotment-yjs.ts`; the `addInitScript` Playwright seeds (homepage/onboarding) clear Yjs IndexedDB on first load via a new `clearYjsIdbOnFirstLoad` helper in `tests/utils/storage.ts`; the path-parity test became a Yjs-path regression test (`allotment-yjs-storage.test.ts`), and `useSyncedStorage.test.ts` became `useCloudSync.test.ts`. Step 4 (Yjs binary as the cloud transport, Supabase BYTEA) remains the next deliberate work.
+
+### Superseded: Phase 1 soak then Step 5 cleanup
 
 The soak window is open. Success criterion is qualitative for the two-user cohort: both real users use the app on the flag for ~3–5 days each, run at least one manual cross-device conflict (edit on phone and laptop, watch the conflict-replace path actually re-hydrate the Yjs doc from cloud), and report no data anomalies. The Yjs binary on each device is the source of truth from this point; the legacy `allotment-unified-data` localStorage key is being mirrored from Yjs and is the rollback floor. Step 5 then deletes `useSyncedStorage`, the legacy branch of every domain-hook method, `useYjsToLegacyMirror`, the `bwp-storage-flag` BroadcastChannel, the legacy localStorage key, and the same-tab broadcast apparatus from PR #369 (the `bonnie:storage-update` CustomEvent, the `instanceId`/`sameTabSeq` bookkeeping, the `recordSavedState`/`recordAdoptedState` helpers, the `recentSavesRef` echo dedup) — every line of that broadcast becomes redundant the day the legacy chain leaves the tree. `serializeToJson` and `decodeDocState` stay forever (rollback + GDPR export + debug). Separate follow-ups worth filing: rename `src/lib/yjs-spike/` → `src/lib/yjs/`, consolidate `src/hooks/allotment/yjs-helpers.ts` with the now-internal `assignDefined` in `allotment-yjs.ts`, and tighten the still-`addInitScript`-pattern Playwright seeds (homepage / onboarding / boost-this-bed) to also clear Yjs IDB if those tests start contaminating each other later.
