--- src/__tests__/hooks/usePersistedStorage.test.ts
@@ -1,640 +1,0 @@
-/**
- * Unit tests for usePersistedStorage hook
- *
- * Focus on flush mechanism and race condition handling
- */
-
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
-import { renderHook, act, waitFor } from '@testing-library/react'
-import { usePersistedStorage } from '@/hooks/usePersistedStorage'
-
-// Mock data type for testing
-interface TestData {
-  version: number
-  value: string
-}
-
-describe('usePersistedStorage - Flush Mechanism', () => {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let mockLoad: any
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let mockSave: any
-
-  beforeEach(() => {
-    mockLoad = vi.fn()
-    mockSave = vi.fn()
-  })
-
-  it('flushSave returns true when save succeeds', async () => {
-    const testData: TestData = { version: 1, value: 'initial' }
-    const updatedData: TestData = { version: 1, value: 'updated' }
-
-    // Initial load returns original data
-    mockLoad.mockReturnValueOnce({ success: true, data: testData })
-    // Save succeeds
-    mockSave.mockReturnValue({ success: true })
-    // Verification load returns the updated data (what we just saved)
-    mockLoad.mockReturnValue({ success: true, data: updatedData })
-
-    const { result } = renderHook(() =>
-      usePersistedStorage<TestData>({
-        storageKey: 'test-key',
-        load: mockLoad,
-        save: mockSave,
-      })
-    )
-
-    // Wait for initial load
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    // Make a change
-    act(() => {
-      result.current.setData(updatedData)
-    })
-
-    // Flush immediately
-    let flushResult: boolean = false
-    await act(async () => {
-      flushResult = await result.current.flushSave()
-    })
-
-    expect(flushResult).toBe(true)
-    expect(mockSave).toHaveBeenCalledWith(updatedData)
-  })
-
-  it('flushSave returns false when save fails', async () => {
-    const testData: TestData = { version: 1, value: 'initial' }
-    mockLoad.mockReturnValue({ success: true, data: testData })
-    mockSave.mockReturnValue({ success: false, error: 'Save failed' })
-
-    const { result } = renderHook(() =>
-      usePersistedStorage<TestData>({
-        storageKey: 'test-key',
-        load: mockLoad,
-        save: mockSave,
-      })
-    )
-
-    // Wait for initial load
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    // Make a change
-    act(() => {
-      result.current.setData({ version: 1, value: 'updated' })
-    })
-
-    // Flush immediately
-    let flushResult: boolean = true
-    await act(async () => {
-      flushResult = await result.current.flushSave()
-    })
-
-    expect(flushResult).toBe(false)
-    expect(result.current.saveStatus).toBe('error')
-  })
-
-  it('flushSave verifies data after write', async () => {
-    const testData: TestData = { version: 1, value: 'initial' }
-    const updatedData: TestData = { version: 1, value: 'updated' }
-
-    // Initial load returns original data
-    mockLoad.mockReturnValueOnce({ success: true, data: testData })
-    // Save succeeds
-    mockSave.mockReturnValue({ success: true })
-    // Verification load returns the updated data
-    mockLoad.mockReturnValueOnce({ success: true, data: updatedData })
-
-    const { result } = renderHook(() =>
-      usePersistedStorage<TestData>({
-        storageKey: 'test-key',
-        load: mockLoad,
-        save: mockSave,
-      })
-    )
-
-    // Wait for initial load
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    // Make a change
-    act(() => {
-      result.current.setData(updatedData)
-    })
-
-    // Flush immediately
-    let flushResult: boolean = false
-    await act(async () => {
-      flushResult = await result.current.flushSave()
-    })
-
-    expect(flushResult).toBe(true)
-    // Should have called load twice: once for initial, once for verification
-    expect(mockLoad).toHaveBeenCalledTimes(2)
-  })
-
-  it('flushSave returns false when verification load fails', async () => {
-    const testData: TestData = { version: 1, value: 'initial' }
-    const updatedData: TestData = { version: 1, value: 'updated' }
-
-    // Initial load returns original data
-    mockLoad.mockReturnValueOnce({ success: true, data: testData })
-    // Save succeeds
-    mockSave.mockReturnValue({ success: true })
-    // Verification load fails (write didn't persist correctly)
-    mockLoad.mockReturnValueOnce({ success: false, error: 'Read failed' })
-
-    const { result } = renderHook(() =>
-      usePersistedStorage<TestData>({
-        storageKey: 'test-key',
-        load: mockLoad,
-        save: mockSave,
-      })
-    )
-
-    // Wait for initial load
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    // Make a change
-    act(() => {
-      result.current.setData(updatedData)
-    })
-
-    // Flush immediately
-    let flushResult: boolean = true
-    await act(async () => {
-      flushResult = await result.current.flushSave()
-    })
-
-    expect(flushResult).toBe(false)
-  })
-
-  it('flushSave returns true when no pending data', async () => {
-    const testData: TestData = { version: 1, value: 'initial' }
-    mockLoad.mockReturnValue({ success: true, data: testData })
-    mockSave.mockReturnValue({ success: true })
-
-    const { result } = renderHook(() =>
-      usePersistedStorage<TestData>({
-        storageKey: 'test-key',
-        load: mockLoad,
-        save: mockSave,
-      })
-    )
-
-    // Wait for initial load
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    // Wait for auto-save to complete (status becomes 'saved' then 'idle' after 2s)
-    await waitFor(() => expect(result.current.saveStatus).toBe('idle'), { timeout: 3000 })
-
-    // Clear any calls from auto-save
-    mockSave.mockClear()
-
-    // Don't make any changes, just flush
-    let flushResult: boolean = false
-    await act(async () => {
-      flushResult = await result.current.flushSave()
-    })
-
-    expect(flushResult).toBe(true)
-    // Should not have called save since there's no pending data after initial auto-save
-    expect(mockSave).not.toHaveBeenCalled()
-  })
-
-  it('flushSave clears timeout when called', async () => {
-    const testData: TestData = { version: 1, value: 'initial' }
-    mockLoad.mockReturnValue({ success: true, data: testData })
-    mockSave.mockReturnValue({ success: true })
-
-    const { result } = renderHook(() =>
-      usePersistedStorage<TestData>({
-        storageKey: 'test-key',
-        load: mockLoad,
-        save: mockSave,
-      })
-    )
-
-    // Wait for initial load
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    // Make a change (this schedules a debounced save)
-    act(() => {
-      result.current.setData({ version: 1, value: 'updated' })
-    })
-
-    expect(result.current.saveStatus).toBe('saving')
-
-    // Flush immediately (should cancel debounced save and save immediately)
-    await act(async () => {
-      await result.current.flushSave()
-    })
-
-    // Should have called save exactly once (from flush, not from debounce)
-    expect(mockSave).toHaveBeenCalledTimes(1)
-  })
-})
-
-describe('usePersistedStorage - Race Condition Handling', () => {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let mockLoad: any
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let mockSave: any
-
-  beforeEach(() => {
-    mockLoad = vi.fn()
-    mockSave = vi.fn()
-  })
-
-  afterEach(() => {
-    // Restore any `vi.spyOn` mocks (e.g. console.warn) so they don't leak
-    // across tests.
-    vi.restoreAllMocks()
-  })
-
-  it('handles concurrent save/import operations via flush', async () => {
-    const testData: TestData = { version: 1, value: 'initial' }
-    const updatedData: TestData = { version: 1, value: 'updated' }
-    const importedData: TestData = { version: 1, value: 'imported' }
-
-    // Initial load returns original data
-    mockLoad.mockReturnValueOnce({ success: true, data: testData })
-    // Save succeeds
-    mockSave.mockReturnValue({ success: true })
-    // Verification load returns the updated data
-    mockLoad.mockReturnValue({ success: true, data: updatedData })
-
-    const { result } = renderHook(() =>
-      usePersistedStorage<TestData>({
-        storageKey: 'test-key',
-        load: mockLoad,
-        save: mockSave,
-      })
-    )
-
-    // Wait for initial load
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    // Make a change (simulating user edit)
-    act(() => {
-      result.current.setData(updatedData)
-    })
-
-    // Flush before import (this is what useDataTransfer does)
-    let flushResult: boolean = false
-    await act(async () => {
-      flushResult = await result.current.flushSave()
-    })
-
-    expect(flushResult).toBe(true)
-    expect(mockSave).toHaveBeenCalledWith(updatedData)
-
-    // Now simulate import by setting new data (after flush completes)
-    mockLoad.mockReturnValue({ success: true, data: importedData })
-
-    act(() => {
-      result.current.setData(importedData)
-    })
-
-    // Data should be the imported data
-    expect(result.current.data).toEqual(importedData)
-
-    // Clear the mock and flush again to verify imported data persists
-    mockSave.mockClear()
-    await act(async () => {
-      await result.current.flushSave()
-    })
-
-    // Should have saved the imported data
-    expect(mockSave).toHaveBeenCalledWith(importedData)
-  })
-
-  it('handles multi-tab sync during import correctly', async () => {
-    const testData: TestData = { version: 1, value: 'initial' }
-    const importedData: TestData = { version: 1, value: 'imported-from-other-tab' }
-
-    // Initial load returns original data
-    mockLoad.mockReturnValueOnce({ success: true, data: testData })
-    mockSave.mockReturnValue({ success: true })
-
-    const { result } = renderHook(() =>
-      usePersistedStorage<TestData>({
-        storageKey: 'test-key',
-        load: mockLoad,
-        save: mockSave,
-      })
-    )
-
-    // Wait for initial load
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    // Simulate storage event from another tab (import happened there)
-    mockLoad.mockReturnValue({ success: true, data: importedData })
-
-    act(() => {
-      // Create and dispatch storage event
-      const storageEvent = new StorageEvent('storage', {
-        key: 'test-key',
-        newValue: JSON.stringify(importedData),
-        oldValue: JSON.stringify(testData),
-      })
-      window.dispatchEvent(storageEvent)
-    })
-
-    // Data should sync to imported data from other tab
-    await waitFor(() => {
-      expect(result.current.data).toEqual(importedData)
-    })
-
-    // Should have called load to sync
-    expect(mockLoad).toHaveBeenCalledTimes(2) // Once for initial, once for storage event
-  })
-
-  it('clears debounce timeout properly when flush is called', async () => {
-    const testData: TestData = { version: 1, value: 'initial' }
-    const updatedData: TestData = { version: 1, value: 'updated' }
-
-    mockLoad.mockReturnValue({ success: true, data: testData })
-    mockSave.mockReturnValue({ success: true })
-
-    const { result } = renderHook(() =>
-      usePersistedStorage<TestData>({
-        storageKey: 'test-key',
-        load: mockLoad,
-        save: mockSave,
-      })
-    )
-
-    // Wait for initial load
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    // Clear any calls from initial load
-    mockSave.mockClear()
-
-    // Make a change (schedules debounced save)
-    act(() => {
-      result.current.setData(updatedData)
-    })
-
-    expect(result.current.saveStatus).toBe('saving')
-
-    // Immediately flush (should cancel debounce and save immediately)
-    await act(async () => {
-      await result.current.flushSave()
-    })
-
-    // Should have called save exactly once (from flush)
-    expect(mockSave).toHaveBeenCalledTimes(1)
-    expect(mockSave).toHaveBeenCalledWith(updatedData)
-
-    // Clear mock and wait for debounce duration to verify no second save
-    mockSave.mockClear()
-
-    // Wait longer than debounce delay
-    await new Promise(resolve => setTimeout(resolve, 1500))
-
-    // Should NOT have called save again (debounce was cancelled)
-    expect(mockSave).not.toHaveBeenCalled()
-  })
-
-  it('propagates same-tab writes to a sibling instance with the same storageKey', async () => {
-    // Repro for the "Turn off Aitor" bug: two usePersistedStorage instances
-    // share a storage key (e.g. Settings page + AitorAuthGate), but the
-    // browser `storage` event does not fire in the writing tab. Verify the
-    // same-tab CustomEvent broadcast wakes up the sibling.
-    const initial: TestData = { version: 1, value: 'initial' }
-    const next: TestData = { version: 1, value: 'aitor-off' }
-
-    // Both instances start with the same initial data.
-    mockLoad.mockReturnValue({ success: true, data: initial })
-    mockSave.mockReturnValue({ success: true })
-
-    const writer = renderHook(() =>
-      usePersistedStorage<TestData>({
-        storageKey: 'shared-key',
-        load: mockLoad,
-        save: mockSave,
-      })
-    )
-    const reader = renderHook(() =>
-      usePersistedStorage<TestData>({
-        storageKey: 'shared-key',
-        load: mockLoad,
-        save: mockSave,
-      })
-    )
-
-    await waitFor(() => expect(writer.result.current.isLoading).toBe(false))
-    await waitFor(() => expect(reader.result.current.isLoading).toBe(false))
-
-    // Writer flips a value (Settings page toggles Aitor off).
-    act(() => {
-      writer.result.current.setData(next)
-    })
-
-    // Wait past the 500ms debounce so the same-tab broadcast fires.
-    await waitFor(
-      () => {
-        expect(reader.result.current.data).toEqual(next)
-      },
-      { timeout: 2000 }
-    )
-
-    // Writer's own listener must not clobber its own state.
-    expect(writer.result.current.data).toEqual(next)
-  })
-
-  it('does not echo same-tab broadcasts back to the originating instance', async () => {
-    // The writer should ignore the event it dispatched itself — otherwise we
-    // risk a render-loop or unnecessary state churn.
-    const initial: TestData = { version: 1, value: 'initial' }
-    const next: TestData = { version: 1, value: 'changed' }
-
-    mockLoad.mockReturnValue({ success: true, data: initial })
-    mockSave.mockReturnValue({ success: true })
-
-    const writer = renderHook(() =>
-      usePersistedStorage<TestData>({
-        storageKey: 'echo-key',
-        load: mockLoad,
-        save: mockSave,
-      })
-    )
-    await waitFor(() => expect(writer.result.current.isLoading).toBe(false))
-
-    act(() => {
-      writer.result.current.setData(next)
-    })
-
-    // After the debounce + broadcast cycle the writer's data should be `next`
-    // (the value it set), not re-loaded from `load` via the broadcast handler.
-    await waitFor(() => expect(writer.result.current.data).toEqual(next))
-
-    // Two load calls total: one initial load. The writer must NOT re-call
-    // `load` in response to its own broadcast.
-    expect(mockLoad).toHaveBeenCalledTimes(1)
-  })
-
-  it('does not re-save data adopted from a cross-tab storage event', async () => {
-    // After a cross-tab `storage` event sets new data, the auto-save effect
-    // would naively treat the new state as a local edit and write the same
-    // value back to localStorage — that write would broadcast a `storage`
-    // event in the other direction and risk a ping-pong. The handler must
-    // seed `latestSerializedRef` so `debouncedSave` skips the redundant save.
-    const initial: TestData = { version: 1, value: 'initial' }
-    const fromOtherTab: TestData = { version: 1, value: 'from-other-tab' }
-
-    mockLoad.mockReturnValueOnce({ success: true, data: initial })
-    mockSave.mockReturnValue({ success: true })
-
-    const { result } = renderHook(() =>
-      usePersistedStorage<TestData>({
-        storageKey: 'cross-tab-key',
-        load: mockLoad,
-        save: mockSave,
-      })
-    )
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    // Clear any saves from initial auto-save settling.
-    mockSave.mockClear()
-
-    // Subsequent loads return the other-tab payload.
-    mockLoad.mockReturnValue({ success: true, data: fromOtherTab })
-
-    // Simulate a `storage` event from another tab.
-    act(() => {
-      window.dispatchEvent(
-        new StorageEvent('storage', {
-          key: 'cross-tab-key',
-          newValue: JSON.stringify(fromOtherTab),
-          oldValue: JSON.stringify(initial),
-        })
-      )
-    })
-
-    await waitFor(() => {
-      expect(result.current.data).toEqual(fromOtherTab)
-    })
-
-    // Give the auto-save debounce + status reset enough time to fire.
-    await new Promise(resolve => setTimeout(resolve, 1500))
-
-    // `save` must NOT have been called for the adopted data — otherwise we'd
-    // be writing back to localStorage exactly what arrived from the other tab.
-    expect(mockSave).not.toHaveBeenCalled()
-  })
-
-  it('warns and ignores same-tab broadcasts whose payload fails validation', async () => {
-    // The same-tab listener defensively re-validates broadcast payloads
-    // before adopting them. If `validate` rejects, the listener should log a
-    // warning and bail without touching the local instance's state.
-    const initial: TestData = { version: 1, value: 'initial' }
-    const invalid: TestData = { version: 1, value: 'invalid' }
-
-    mockLoad.mockReturnValue({ success: true, data: initial })
-    mockSave.mockReturnValue({ success: true })
-
-    const validate = vi.fn().mockReturnValue({
-      success: false,
-      error: 'bad shape',
-    })
-    const warnSpy = vi.spyOn(console, 'warn').mockImplementation(() => {})
-
-    const { result } = renderHook(() =>
-      usePersistedStorage<TestData>({
-        storageKey: 'validate-reject-key',
-        load: mockLoad,
-        save: mockSave,
-        validate,
-      })
-    )
-
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-    expect(result.current.data).toEqual(initial)
-
-    // Dispatch a same-tab broadcast that passes the storageKey / instanceId /
-    // seq gates but should be rejected by `validate`. Use a foreign
-    // `instanceId` so the listener does not skip it as its own echo, and a
-    // large `seq` so the last-writer-wins guard lets it through.
-    act(() => {
-      window.dispatchEvent(
-        new CustomEvent('bonnie:storage-update', {
-          detail: {
-            storageKey: 'validate-reject-key',
-            instanceId: 'foreign-instance',
-            seq: Number.MAX_SAFE_INTEGER,
-            data: invalid,
-            serialized: JSON.stringify(invalid),
-          },
-        })
-      )
-    })
-
-    // Validator must have been consulted, and it rejected the payload, so the
-    // local instance's `data` must remain the initial value (setData not
-    // called) and a warning must have been logged exactly once.
-    expect(validate).toHaveBeenCalledWith(invalid)
-    expect(result.current.data).toEqual(initial)
-    expect(warnSpy).toHaveBeenCalledTimes(1)
-    expect(warnSpy).toHaveBeenCalledWith(
-      'Ignoring invalid same-tab broadcast:',
-      'bad shape'
-    )
-  })
-
-  it('handles pending data correctly - waits for flush before import', async () => {
-    const testData: TestData = { version: 1, value: 'initial' }
-    const pendingData: TestData = { version: 1, value: 'pending-edit' }
-    const importData: TestData = { version: 1, value: 'import' }
-
-    // Initial load
-    mockLoad.mockReturnValueOnce({ success: true, data: testData })
-    mockSave.mockReturnValue({ success: true })
-    // Verification load after save
-    mockLoad.mockReturnValue({ success: true, data: pendingData })
-
-    const { result } = renderHook(() =>
-      usePersistedStorage<TestData>({
-        storageKey: 'test-key',
-        load: mockLoad,
-        save: mockSave,
-      })
-    )
-
-    // Wait for initial load
-    await waitFor(() => expect(result.current.isLoading).toBe(false))
-
-    // Make an edit (creates pending data)
-    act(() => {
-      result.current.setData(pendingData)
-    })
-
-    expect(result.current.saveStatus).toBe('saving')
-
-    // Simulate import workflow - flush first
-    let flushSuccess = false
-    await act(async () => {
-      flushSuccess = await result.current.flushSave()
-    })
-
-    expect(flushSuccess).toBe(true)
-    expect(mockSave).toHaveBeenCalledWith(pendingData)
-
-    // Now it's safe to import
-    mockLoad.mockReturnValue({ success: true, data: importData })
-
-    act(() => {
-      result.current.setData(importData)
-    })
-
-    expect(result.current.data).toEqual(importData)
-
-    // Verify import data can be saved
-    mockSave.mockClear()
-    await act(async () => {
-      await result.current.flushSave()
-    })
-
-    expect(mockSave).toHaveBeenCalledWith(importData)
-  })
-})
