+++ src/hooks/allotment/useAllotmentCareLogs.ts
@@ -4,9 +4,8 @@
  * Care logs and harvest tracking for permanent areas.
  * Handles care log CRUD and harvest totals.
  *
- * Two-branch methods (ADR 027 Step 3, PR-B): see useAllotmentAreas for
- * the convention. The Yjs branch mutates the targeted `AreaSeason`'s
- * `careLogs` array in place.
+ * Writes go through `mutate(fn)` against the SyncedStore proxy (ADR 027;
+ * legacy `setData` branch removed in Step 5).
  */
 
 'use client'
@@ -19,24 +18,18 @@ import {
   AreaSeason,
 } from '@/types/unified-allotment'
 import {
-  addCareLogEntry as storageAddCareLogEntry,
-  updateCareLogEntry as storageUpdateCareLogEntry,
-  removeCareLogEntry as storageRemoveCareLogEntry,
   getCareLogsForArea,
   getAllCareLogsForArea,
-  logHarvest as storageLogHarvest,
   getHarvestTotal,
 } from '@/services/allotment-storage'
 import { generateId } from '@/lib/utils'
-import { USE_YJS_STORAGE } from '@/config/release-visibility'
 import type { MutateFn } from './useAllotmentData'
-import { assignDefined, withoutUndefined } from './yjs-helpers'
+import { assignDefined, withoutUndefined } from '@/lib/yjs/allotment-yjs'
 
 // ============ HOOK TYPES ============
 
 export interface UseAllotmentCareLogsProps {
   data: AllotmentData | null
-  setData: (data: AllotmentData | ((prev: AllotmentData | null) => AllotmentData | null)) => void
   mutate: MutateFn
   selectedYear: number
 }
@@ -55,7 +48,6 @@ export interface UseAllotmentCareLogsReturn {
 
 export function useAllotmentCareLogs({
   data,
-  setData,
   mutate,
   selectedYear,
 }: UseAllotmentCareLogsProps): UseAllotmentCareLogsReturn {
@@ -63,41 +55,35 @@ export function useAllotmentCareLogs({
   const addCareLogData = useCallback((areaId: string, entry: NewCareLogEntry): string => {
     if (!data) return ''
 
-    if (USE_YJS_STORAGE) {
-      // ID has to be computed up front so we can return it synchronously
-      // — the mutate callback runs inside the Yjs transaction but the
-      // ID is generated by `generateId` outside it (the legacy path
-      // does the same, computing the ID inside `addCareLogEntry`).
-      const id = generateId()
-      mutate(store => {
-        const season = store.seasons.find(s => s.year === selectedYear)
-        if (!season) return
-        const now = new Date().toISOString()
+    // ID has to be computed up front so we can return it synchronously
+    // — the mutate callback runs inside the Yjs transaction but the
+    // ID is generated by `generateId` outside it (the legacy path
+    // does the same, computing the ID inside `addCareLogEntry`).
+    const id = generateId()
+    mutate(store => {
+      const season = store.seasons.find(s => s.year === selectedYear)
+      if (!season) return
+      const now = new Date().toISOString()
 
-        let areaSeason = season.areas.find(a => a.areaId === areaId)
-        if (!areaSeason) {
-          const fresh: AreaSeason = withoutUndefined({
-            areaId,
-            plantings: [],
-            careLogs: [],
-          })
-          season.areas.push(fresh)
-          areaSeason = season.areas[season.areas.length - 1]
-        }
+      let areaSeason = season.areas.find(a => a.areaId === areaId)
+      if (!areaSeason) {
+        const fresh: AreaSeason = withoutUndefined({
+          areaId,
+          plantings: [],
+          careLogs: [],
+        })
+        season.areas.push(fresh)
+        areaSeason = season.areas[season.areas.length - 1]
+      }
 
-        if (!areaSeason.careLogs) areaSeason.careLogs = []
-        const newEntry: CareLogEntry = withoutUndefined({ ...entry, id })
-        areaSeason.careLogs.push(newEntry)
-        season.updatedAt = now
-        store.meta.updatedAt = now
-      })
-      return id
-    }
-
-    const result = storageAddCareLogEntry(data, selectedYear, areaId, entry)
-    setData(result.data)
-    return result.entryId
-  }, [data, selectedYear, setData, mutate])
+      if (!areaSeason.careLogs) areaSeason.careLogs = []
+      const newEntry: CareLogEntry = withoutUndefined({ ...entry, id })
+      areaSeason.careLogs.push(newEntry)
+      season.updatedAt = now
+      store.meta.updatedAt = now
+    })
+    return id
+  }, [data, selectedYear, mutate])
 
   const updateCareLogData = useCallback((
     areaId: string,
@@ -106,46 +92,36 @@ export function useAllotmentCareLogs({
   ) => {
     if (!data) return
 
-    if (USE_YJS_STORAGE) {
-      mutate(store => {
-        const season = store.seasons.find(s => s.year === selectedYear)
-        if (!season) return
-        const areaSeason = season.areas.find(a => a.areaId === areaId)
-        if (!areaSeason?.careLogs) return
-        const log = areaSeason.careLogs.find(l => l.id === entryId)
-        if (!log) return
-        const now = new Date().toISOString()
-        assignDefined(log as unknown as Record<string, unknown>, updates as Record<string, unknown>)
-        season.updatedAt = now
-        store.meta.updatedAt = now
-      })
-      return
-    }
-
-    setData(storageUpdateCareLogEntry(data, selectedYear, areaId, entryId, updates))
-  }, [data, selectedYear, setData, mutate])
+    mutate(store => {
+      const season = store.seasons.find(s => s.year === selectedYear)
+      if (!season) return
+      const areaSeason = season.areas.find(a => a.areaId === areaId)
+      if (!areaSeason?.careLogs) return
+      const log = areaSeason.careLogs.find(l => l.id === entryId)
+      if (!log) return
+      const now = new Date().toISOString()
+      assignDefined(log as unknown as Record<string, unknown>, updates as Record<string, unknown>)
+      season.updatedAt = now
+      store.meta.updatedAt = now
+    })
+  }, [data, selectedYear, mutate])
 
   const removeCareLogData = useCallback((areaId: string, entryId: string) => {
     if (!data) return
 
-    if (USE_YJS_STORAGE) {
-      mutate(store => {
-        const season = store.seasons.find(s => s.year === selectedYear)
-        if (!season) return
-        const areaSeason = season.areas.find(a => a.areaId === areaId)
-        if (!areaSeason?.careLogs) return
-        const idx = areaSeason.careLogs.findIndex(l => l.id === entryId)
-        if (idx === -1) return
-        areaSeason.careLogs.splice(idx, 1)
-        const now = new Date().toISOString()
-        season.updatedAt = now
-        store.meta.updatedAt = now
-      })
-      return
-    }
-
-    setData(storageRemoveCareLogEntry(data, selectedYear, areaId, entryId))
-  }, [data, selectedYear, setData, mutate])
+    mutate(store => {
+      const season = store.seasons.find(s => s.year === selectedYear)
+      if (!season) return
+      const areaSeason = season.areas.find(a => a.areaId === areaId)
+      if (!areaSeason?.careLogs) return
+      const idx = areaSeason.careLogs.findIndex(l => l.id === entryId)
+      if (idx === -1) return
+      areaSeason.careLogs.splice(idx, 1)
+      const now = new Date().toISOString()
+      season.updatedAt = now
+      store.meta.updatedAt = now
+    })
+  }, [data, selectedYear, mutate])
 
   const getCareLogsData = useCallback((areaId: string): CareLogEntry[] => {
     if (!data) return []
@@ -165,44 +141,36 @@ export function useAllotmentCareLogs({
   ): string => {
     if (!data) return ''
 
-    if (USE_YJS_STORAGE) {
-      // logHarvest is a convenience wrapper around addCareLogEntry on
-      // the legacy path. Mirror that by routing through the same Yjs
-      // branch — keeps the two implementations textually close.
-      const id = generateId()
-      mutate(store => {
-        const season = store.seasons.find(s => s.year === selectedYear)
-        if (!season) return
-        const now = new Date().toISOString()
-        let areaSeason = season.areas.find(a => a.areaId === areaId)
-        if (!areaSeason) {
-          const fresh: AreaSeason = withoutUndefined({
-            areaId,
-            plantings: [],
-            careLogs: [],
-          })
-          season.areas.push(fresh)
-          areaSeason = season.areas[season.areas.length - 1]
-        }
-        if (!areaSeason.careLogs) areaSeason.careLogs = []
-        const entry: CareLogEntry = withoutUndefined<CareLogEntry>({
-          id,
-          type: 'harvest',
-          date: date || new Date().toISOString().split('T')[0],
-          quantity,
-          unit,
+    // logHarvest is a convenience wrapper around addCareLogEntry.
+    const id = generateId()
+    mutate(store => {
+      const season = store.seasons.find(s => s.year === selectedYear)
+      if (!season) return
+      const now = new Date().toISOString()
+      let areaSeason = season.areas.find(a => a.areaId === areaId)
+      if (!areaSeason) {
+        const fresh: AreaSeason = withoutUndefined({
+          areaId,
+          plantings: [],
+          careLogs: [],
         })
-        areaSeason.careLogs.push(entry)
-        season.updatedAt = now
-        store.meta.updatedAt = now
+        season.areas.push(fresh)
+        areaSeason = season.areas[season.areas.length - 1]
+      }
+      if (!areaSeason.careLogs) areaSeason.careLogs = []
+      const entry: CareLogEntry = withoutUndefined<CareLogEntry>({
+        id,
+        type: 'harvest',
+        date: date || new Date().toISOString().split('T')[0],
+        quantity,
+        unit,
       })
-      return id
-    }
-
-    const result = storageLogHarvest(data, selectedYear, areaId, quantity, unit, date)
-    setData(result.data)
-    return result.entryId
-  }, [data, selectedYear, setData, mutate])
+      areaSeason.careLogs.push(entry)
+      season.updatedAt = now
+      store.meta.updatedAt = now
+    })
+    return id
+  }, [data, selectedYear, mutate])
 
   const getHarvestTotalData = useCallback((areaId: string): { quantity: number; unit: string } | null => {
     if (!data) return null
