+++ src/hooks/allotment/useAllotmentCustomTasks.ts
@@ -3,8 +3,8 @@
  *
  * CRUD operations for free-form custom tasks on the Today dashboard.
  *
- * Two-branch methods (ADR 027 Step 3, PR-B): see useAllotmentAreas for
- * the convention. The Yjs branch mutates `store.customTasks` in place.
+ * Writes go through `mutate(fn)` against the SyncedStore proxy (ADR 027;
+ * legacy `setData` branch removed in Step 5).
  */
 
 'use client'
@@ -17,21 +17,15 @@ import {
 } from '@/types/unified-allotment'
 import {
   getCustomTasks as storageGetCustomTasks,
-  addCustomTask as storageAddCustomTask,
-  toggleCustomTask as storageToggleCustomTask,
-  updateCustomTask as storageUpdateCustomTask,
-  removeCustomTask as storageRemoveCustomTask,
 } from '@/services/allotment-storage'
 import { generateId } from '@/lib/utils'
-import { USE_YJS_STORAGE } from '@/config/release-visibility'
 import type { MutateFn } from './useAllotmentData'
-import { withoutUndefined } from './yjs-helpers'
+import { withoutUndefined } from '@/lib/yjs/allotment-yjs'
 
 // ============ HOOK TYPES ============
 
 export interface UseAllotmentCustomTasksProps {
   data: AllotmentData | null
-  setData: (data: AllotmentData | ((prev: AllotmentData | null) => AllotmentData | null)) => void
   mutate: MutateFn
 }
 
@@ -47,7 +41,6 @@ export interface UseAllotmentCustomTasksReturn {
 
 export function useAllotmentCustomTasks({
   data,
-  setData,
   mutate,
 }: UseAllotmentCustomTasksProps): UseAllotmentCustomTasksReturn {
 
@@ -59,77 +52,57 @@ export function useAllotmentCustomTasks({
   const addTask = useCallback((task: NewCustomTask) => {
     if (!data) return
 
-    if (USE_YJS_STORAGE) {
-      mutate(store => {
-        const newTask: CustomTask = withoutUndefined({
-          id: generateId('custom-task'),
-          description: task.description,
-          completed: false,
-          createdAt: new Date().toISOString(),
-        })
-        // Legacy storage prepends new tasks (newest first); mirror that
-        // ordering so the parity snapshot matches.
-        store.customTasks.unshift(newTask)
+    mutate(store => {
+      const newTask: CustomTask = withoutUndefined({
+        id: generateId('custom-task'),
+        description: task.description,
+        completed: false,
+        createdAt: new Date().toISOString(),
       })
-      return
-    }
-
-    setData(storageAddCustomTask(data, task))
-  }, [data, setData, mutate])
+      // Legacy storage prepends new tasks (newest first); mirror that
+      // ordering so the parity snapshot matches.
+      store.customTasks.unshift(newTask)
+    })
+  }, [data, mutate])
 
   const toggleTask = useCallback((taskId: string) => {
     if (!data) return
 
-    if (USE_YJS_STORAGE) {
-      mutate(store => {
-        const t = store.customTasks.find(x => x.id === taskId)
-        if (!t) return
-        const nextCompleted = !t.completed
-        t.completed = nextCompleted
-        if (nextCompleted) {
-          t.completedAt = new Date().toISOString()
-        } else {
-          // Legacy assigns `undefined` here; on Yjs delete the field so
-          // the serialized snapshot omits it (matches the JSON.stringify
-          // behaviour of the legacy path).
-          delete (t as Partial<CustomTask>).completedAt
-        }
-      })
-      return
-    }
-
-    setData(storageToggleCustomTask(data, taskId))
-  }, [data, setData, mutate])
+    mutate(store => {
+      const t = store.customTasks.find(x => x.id === taskId)
+      if (!t) return
+      const nextCompleted = !t.completed
+      t.completed = nextCompleted
+      if (nextCompleted) {
+        t.completedAt = new Date().toISOString()
+      } else {
+        // Legacy assigns `undefined` here; on Yjs delete the field so
+        // the serialized snapshot omits it (matches the JSON.stringify
+        // behaviour of the legacy path).
+        delete (t as Partial<CustomTask>).completedAt
+      }
+    })
+  }, [data, mutate])
 
   const updateTask = useCallback((taskId: string, description: string) => {
     if (!data) return
 
-    if (USE_YJS_STORAGE) {
-      mutate(store => {
-        const t = store.customTasks.find(x => x.id === taskId)
-        if (!t) return
-        t.description = description
-      })
-      return
-    }
-
-    setData(storageUpdateCustomTask(data, taskId, description))
-  }, [data, setData, mutate])
+    mutate(store => {
+      const t = store.customTasks.find(x => x.id === taskId)
+      if (!t) return
+      t.description = description
+    })
+  }, [data, mutate])
 
   const removeTask = useCallback((taskId: string) => {
     if (!data) return
 
-    if (USE_YJS_STORAGE) {
-      mutate(store => {
-        const idx = store.customTasks.findIndex(t => t.id === taskId)
-        if (idx === -1) return
-        store.customTasks.splice(idx, 1)
-      })
-      return
-    }
-
-    setData(storageRemoveCustomTask(data, taskId))
-  }, [data, setData, mutate])
+    mutate(store => {
+      const idx = store.customTasks.findIndex(t => t.id === taskId)
+      if (idx === -1) return
+      store.customTasks.splice(idx, 1)
+    })
+  }, [data, mutate])
 
   return {
     getCustomTasks: getCustomTasksData,
