+++ src/hooks/allotment/useAllotmentPlantings.ts
@@ -4,9 +4,9 @@
  * Planting CRUD with variety auto-sync.
  * Handles plantings, area seasons, season operations, and rotation tracking.
  *
- * Two-branch methods (ADR 027 Step 3, PR-B): see useAllotmentAreas for
- * the convention. `addPlanting` and `addPlantings` are the
- * cross-collection methods called out in the spec — they touch
+ * Writes go through `mutate(fn)` against the SyncedStore proxy (ADR 027;
+ * the legacy `setData` branch was removed in Step 5). `addPlanting` and
+ * `addPlantings` are cross-collection — they touch
  * `store.seasons[...].areas[...].plantings` AND `store.varieties` in
  * one Yjs transaction.
  */
@@ -28,15 +28,6 @@ import {
 import { PhysicalBedId, RotationGroup } from '@/types/garden-planner'
 import {
   getAreaSeason as storageGetAreaSeason,
-  addPlanting as storageAddPlanting,
-  addPlantings as storageAddPlantings,
-  updatePlanting as storageUpdatePlanting,
-  removePlanting as storageRemovePlanting,
-  addSeason,
-  removeSeason,
-  updateSeason,
-  updateAreaRotationGroup as storageUpdateAreaRotationGroup,
-  updateAreaSeasonPosition as storageUpdateAreaSeasonPosition,
   getPlantingsForArea as storageGetPlantingsForArea,
   getRotationBeds as storageGetRotationBeds,
   getAreasByKind as storageGetAreasByKind,
@@ -47,15 +38,13 @@ import { generateId } from '@/lib/utils/id'
 import { normalizeVarietyName } from '@/lib/variety-queries'
 import { inferStatusFromDates } from '@/lib/planting-utils'
 import { getNextRotationGroup } from '@/lib/rotation'
-import { USE_YJS_STORAGE } from '@/config/release-visibility'
 import type { MutateFn } from './useAllotmentData'
-import { assignDefined, withoutUndefined } from './yjs-helpers'
+import { assignDefined, withoutUndefined } from '@/lib/yjs/allotment-yjs'
 
 // ============ HOOK TYPES ============
 
 export interface UseAllotmentPlantingsProps {
   data: AllotmentData | null
-  setData: (data: AllotmentData | ((prev: AllotmentData | null) => AllotmentData | null)) => void
   mutate: MutateFn
   selectedYear: number
   setSelectedYear?: (year: number) => void
@@ -92,7 +81,6 @@ export interface UseAllotmentPlantingsReturn {
 
 export function useAllotmentPlantings({
   data,
-  setData,
   mutate,
   selectedYear,
   setSelectedYear,
@@ -103,37 +91,89 @@ export function useAllotmentPlantings({
   const addPlanting = useCallback((bedId: PhysicalBedId, planting: NewPlanting) => {
     if (!data) return
 
-    if (USE_YJS_STORAGE) {
-      // Cross-collection: writes plantings AND varieties in one
-      // transaction so a single Yjs update emits and both observers see
-      // a consistent state.
-      const plantingId = generateId('planting')
-      mutate(store => {
-        const season = store.seasons.find(s => s.year === selectedYear)
-        if (!season) return
-        const now = new Date().toISOString()
+    // Cross-collection: writes plantings AND varieties in one
+    // transaction so a single Yjs update emits and both observers see
+    // a consistent state.
+    const plantingId = generateId('planting')
+    mutate(store => {
+      const season = store.seasons.find(s => s.year === selectedYear)
+      if (!season) return
+      const now = new Date().toISOString()
 
-        // Ensure AreaSeason exists.
-        let areaSeason = season.areas.find(a => a.areaId === bedId)
-        if (!areaSeason) {
-          const fresh: AreaSeason = withoutUndefined({
-            areaId: bedId,
-            plantings: [],
+      // Ensure AreaSeason exists.
+      let areaSeason = season.areas.find(a => a.areaId === bedId)
+      if (!areaSeason) {
+        const fresh: AreaSeason = withoutUndefined({
+          areaId: bedId,
+          plantings: [],
+        })
+        season.areas.push(fresh)
+        areaSeason = season.areas[season.areas.length - 1]
+      }
+
+      const newPlanting: Planting = withoutUndefined({
+        ...planting,
+        id: plantingId,
+      })
+      areaSeason.plantings.push(newPlanting)
+      season.updatedAt = now
+
+      // Variety auto-sync: if the planting has a variety name, find
+      // or create the matching StoredVariety so the seed inventory
+      // stays in step with what's actually planted.
+      if (planting.varietyName) {
+        const normalizedName = normalizeVarietyName(planting.varietyName)
+        const existing = store.varieties.find(v =>
+          v.plantId === planting.plantId &&
+          normalizeVarietyName(v.name) === normalizedName,
+        )
+
+        if (!existing) {
+          const newVariety: StoredVariety = withoutUndefined({
+            id: generateId('variety'),
+            plantId: planting.plantId,
+            name: planting.varietyName,
+            notes: '(Auto-created from planting)',
+            seedsByYear: { [selectedYear]: 'none' as const },
+            isArchived: false,
           })
-          season.areas.push(fresh)
-          areaSeason = season.areas[season.areas.length - 1]
+          store.varieties.push(newVariety)
+        } else if (!(selectedYear in (existing.seedsByYear || {}))) {
+          existing.seedsByYear = { ...existing.seedsByYear, [selectedYear]: 'none' }
         }
+      }
+    })
+  }, [data, selectedYear, mutate])
 
+  const addPlantings = useCallback((bedId: PhysicalBedId, plantings: NewPlanting[]) => {
+    if (!data || plantings.length === 0) return
+
+    // Cross-collection: same shape as addPlanting but bulk. Generate
+    // all IDs up front so the parity test stays deterministic under a
+    // mocked `generateId`.
+    const plantingIds = plantings.map(() => generateId('planting'))
+    mutate(store => {
+      const season = store.seasons.find(s => s.year === selectedYear)
+      if (!season) return
+      const now = new Date().toISOString()
+
+      let areaSeason = season.areas.find(a => a.areaId === bedId)
+      if (!areaSeason) {
+        const fresh: AreaSeason = withoutUndefined({
+          areaId: bedId,
+          plantings: [],
+        })
+        season.areas.push(fresh)
+        areaSeason = season.areas[season.areas.length - 1]
+      }
+
+      plantings.forEach((planting, i) => {
         const newPlanting: Planting = withoutUndefined({
           ...planting,
-          id: plantingId,
+          id: plantingIds[i],
         })
-        areaSeason.plantings.push(newPlanting)
-        season.updatedAt = now
+        areaSeason!.plantings.push(newPlanting)
 
-        // Variety auto-sync: if the planting has a variety name, find
-        // or create the matching StoredVariety so the seed inventory
-        // stays in step with what's actually planted.
         if (planting.varietyName) {
           const normalizedName = normalizeVarietyName(planting.varietyName)
           const existing = store.varieties.find(v =>
@@ -156,205 +196,48 @@ export function useAllotmentPlantings({
           }
         }
       })
-      return
-    }
 
-    // Add planting to allotment storage
-    let updatedData = storageAddPlanting(data, selectedYear, bedId, planting)
-
-    // Inline variety sync (no separate storage)
-    if (planting.varietyName) {
-      const normalizedName = normalizeVarietyName(planting.varietyName)
-
-      const variety = updatedData.varieties.find(v =>
-        v.plantId === planting.plantId &&
-        normalizeVarietyName(v.name) === normalizedName
-      )
-
-      if (!variety) {
-        // Auto-create variety with seed status for this year
-        updatedData = {
-          ...updatedData,
-          varieties: [
-            ...updatedData.varieties,
-            {
-              id: generateId('variety'),
-              plantId: planting.plantId,
-              name: planting.varietyName,
-              notes: '(Auto-created from planting)',
-              seedsByYear: { [selectedYear]: 'none' },
-              isArchived: false
-            }
-          ]
-        }
-      } else if (!(selectedYear in (variety.seedsByYear || {}))) {
-        // Add year to existing variety's seedsByYear if not already tracked
-        updatedData = {
-          ...updatedData,
-          varieties: updatedData.varieties.map(v =>
-            v.id === variety.id
-              ? { ...v, seedsByYear: { ...v.seedsByYear, [selectedYear]: 'none' } }
-              : v
-          )
-        }
-      }
-    }
-
-    setData(updatedData)
-  }, [data, selectedYear, setData, mutate])
-
-  const addPlantings = useCallback((bedId: PhysicalBedId, plantings: NewPlanting[]) => {
-    if (!data || plantings.length === 0) return
-
-    if (USE_YJS_STORAGE) {
-      // Cross-collection: same shape as addPlanting but bulk. Generate
-      // all IDs up front so both branches use the same number of
-      // generateId calls in the same order — keeps the parity test
-      // deterministic under a mocked `generateId`.
-      const plantingIds = plantings.map(() => generateId('planting'))
-      mutate(store => {
-        const season = store.seasons.find(s => s.year === selectedYear)
-        if (!season) return
-        const now = new Date().toISOString()
-
-        let areaSeason = season.areas.find(a => a.areaId === bedId)
-        if (!areaSeason) {
-          const fresh: AreaSeason = withoutUndefined({
-            areaId: bedId,
-            plantings: [],
-          })
-          season.areas.push(fresh)
-          areaSeason = season.areas[season.areas.length - 1]
-        }
-
-        plantings.forEach((planting, i) => {
-          const newPlanting: Planting = withoutUndefined({
-            ...planting,
-            id: plantingIds[i],
-          })
-          areaSeason!.plantings.push(newPlanting)
-
-          if (planting.varietyName) {
-            const normalizedName = normalizeVarietyName(planting.varietyName)
-            const existing = store.varieties.find(v =>
-              v.plantId === planting.plantId &&
-              normalizeVarietyName(v.name) === normalizedName,
-            )
-
-            if (!existing) {
-              const newVariety: StoredVariety = withoutUndefined({
-                id: generateId('variety'),
-                plantId: planting.plantId,
-                name: planting.varietyName,
-                notes: '(Auto-created from planting)',
-                seedsByYear: { [selectedYear]: 'none' as const },
-                isArchived: false,
-              })
-              store.varieties.push(newVariety)
-            } else if (!(selectedYear in (existing.seedsByYear || {}))) {
-              existing.seedsByYear = { ...existing.seedsByYear, [selectedYear]: 'none' }
-            }
-          }
-        })
-
-        season.updatedAt = now
-      })
-      return
-    }
-
-    // Add all plantings in a single state update
-    let updatedData = storageAddPlantings(data, selectedYear, bedId, plantings)
-
-    // Sync each planting's variety (inline, no separate storage)
-    plantings.forEach(planting => {
-      if (planting.varietyName) {
-        const normalizedName = normalizeVarietyName(planting.varietyName)
-
-        const variety = updatedData.varieties.find(v =>
-          v.plantId === planting.plantId &&
-          normalizeVarietyName(v.name) === normalizedName
-        )
-
-        if (!variety) {
-          // Auto-create variety with seed status for this year
-          updatedData = {
-            ...updatedData,
-            varieties: [
-              ...updatedData.varieties,
-              {
-                id: generateId('variety'),
-                plantId: planting.plantId,
-                name: planting.varietyName,
-                notes: '(Auto-created from planting)',
-                seedsByYear: { [selectedYear]: 'none' },
-                isArchived: false
-              }
-            ]
-          }
-        } else if (!(selectedYear in (variety.seedsByYear || {}))) {
-          // Add year to existing variety's seedsByYear if not already tracked
-          updatedData = {
-            ...updatedData,
-            varieties: updatedData.varieties.map(v =>
-              v.id === variety.id
-                ? { ...v, seedsByYear: { ...v.seedsByYear, [selectedYear]: 'none' } }
-                : v
-            )
-          }
-        }
-      }
+      season.updatedAt = now
     })
-
-    setData(updatedData)
-  }, [data, selectedYear, setData, mutate])
+  }, [data, selectedYear, mutate])
 
   const updatePlanting = useCallback((bedId: PhysicalBedId, plantingId: string, updates: PlantingUpdate) => {
     if (!data) return
 
-    if (USE_YJS_STORAGE) {
-      mutate(store => {
-        const season = store.seasons.find(s => s.year === selectedYear)
-        if (!season) return
-        const areaSeason = season.areas.find(a => a.areaId === bedId)
-        if (!areaSeason) return
-        const p = areaSeason.plantings.find(x => x.id === plantingId)
-        if (!p) return
+    mutate(store => {
+      const season = store.seasons.find(s => s.year === selectedYear)
+      if (!season) return
+      const areaSeason = season.areas.find(a => a.areaId === bedId)
+      if (!areaSeason) return
+      const p = areaSeason.plantings.find(x => x.id === plantingId)
+      if (!p) return
 
-        const datesChanged =
-          'sowDate' in updates || 'transplantDate' in updates || 'actualHarvestEnd' in updates
-        const statusExplicitlySet = 'status' in updates
+      const datesChanged =
+        'sowDate' in updates || 'transplantDate' in updates || 'actualHarvestEnd' in updates
+      const statusExplicitlySet = 'status' in updates
 
-        assignDefined(p as unknown as Record<string, unknown>, updates as Record<string, unknown>)
-        if (datesChanged && !statusExplicitlySet && p.status !== 'removed') {
-          p.status = inferStatusFromDates(p)
-        }
-        season.updatedAt = new Date().toISOString()
-      })
-      return
-    }
-
-    setData(storageUpdatePlanting(data, selectedYear, bedId, plantingId, updates))
-  }, [data, selectedYear, setData, mutate])
+      assignDefined(p as unknown as Record<string, unknown>, updates as Record<string, unknown>)
+      if (datesChanged && !statusExplicitlySet && p.status !== 'removed') {
+        p.status = inferStatusFromDates(p)
+      }
+      season.updatedAt = new Date().toISOString()
+    })
+  }, [data, selectedYear, mutate])
 
   const removePlanting = useCallback((bedId: PhysicalBedId, plantingId: string) => {
     if (!data) return
 
-    if (USE_YJS_STORAGE) {
-      mutate(store => {
-        const season = store.seasons.find(s => s.year === selectedYear)
-        if (!season) return
-        const areaSeason = season.areas.find(a => a.areaId === bedId)
-        if (!areaSeason) return
-        const idx = areaSeason.plantings.findIndex(p => p.id === plantingId)
-        if (idx === -1) return
-        areaSeason.plantings.splice(idx, 1)
-        season.updatedAt = new Date().toISOString()
-      })
-      return
-    }
-
-    setData(storageRemovePlanting(data, selectedYear, bedId, plantingId))
-  }, [data, selectedYear, setData, mutate])
+    mutate(store => {
+      const season = store.seasons.find(s => s.year === selectedYear)
+      if (!season) return
+      const areaSeason = season.areas.find(a => a.areaId === bedId)
+      if (!areaSeason) return
+      const idx = areaSeason.plantings.findIndex(p => p.id === plantingId)
+      if (idx === -1) return
+      areaSeason.plantings.splice(idx, 1)
+      season.updatedAt = new Date().toISOString()
+    })
+  }, [data, selectedYear, mutate])
 
   const getPlantings = useCallback((areaId: string) => {
     if (!data) return []
@@ -371,154 +254,122 @@ export function useAllotmentPlantings({
   const updateRotationGroup = useCallback((areaId: string, group: RotationGroup) => {
     if (!data) return
 
-    if (USE_YJS_STORAGE) {
-      mutate(store => {
-        const season = store.seasons.find(s => s.year === selectedYear)
-        if (!season) return
-        const existing = season.areas.find(a => a.areaId === areaId)
-        if (existing) {
-          existing.rotationGroup = group
-        } else {
-          const fresh: AreaSeason = withoutUndefined({
-            areaId,
-            rotationGroup: group,
-            plantings: [],
-          })
-          season.areas.push(fresh)
-        }
-        season.updatedAt = new Date().toISOString()
-      })
-      return
-    }
-
-    setData(storageUpdateAreaRotationGroup(data, selectedYear, areaId, group))
-  }, [data, selectedYear, setData, mutate])
+    mutate(store => {
+      const season = store.seasons.find(s => s.year === selectedYear)
+      if (!season) return
+      const existing = season.areas.find(a => a.areaId === areaId)
+      if (existing) {
+        existing.rotationGroup = group
+      } else {
+        const fresh: AreaSeason = withoutUndefined({
+          areaId,
+          rotationGroup: group,
+          plantings: [],
+        })
+        season.areas.push(fresh)
+      }
+      season.updatedAt = new Date().toISOString()
+    })
+  }, [data, selectedYear, mutate])
 
   const updateAreaSeasonPositionFn = useCallback((areaId: string, position: GridPosition) => {
-    if (USE_YJS_STORAGE) {
-      mutate(store => {
-        const season = store.seasons.find(s => s.year === selectedYear)
-        if (!season) return
-        const existing = season.areas.find(a => a.areaId === areaId)
-        if (existing) {
-          existing.gridPosition = position
-        } else {
-          const fresh: AreaSeason = withoutUndefined({
-            areaId,
-            plantings: [],
-            gridPosition: position,
-          })
-          season.areas.push(fresh)
-        }
-        season.updatedAt = new Date().toISOString()
-      })
-      return
-    }
-
-    setData(prevData => {
-      if (!prevData) return prevData
-      return storageUpdateAreaSeasonPosition(prevData, selectedYear, areaId, position)
+    mutate(store => {
+      const season = store.seasons.find(s => s.year === selectedYear)
+      if (!season) return
+      const existing = season.areas.find(a => a.areaId === areaId)
+      if (existing) {
+        existing.gridPosition = position
+      } else {
+        const fresh: AreaSeason = withoutUndefined({
+          areaId,
+          plantings: [],
+          gridPosition: position,
+        })
+        season.areas.push(fresh)
+      }
+      season.updatedAt = new Date().toISOString()
     })
-  }, [selectedYear, setData, mutate])
+  }, [selectedYear, mutate])
 
   // ============ SEASON OPERATIONS ============
 
   const createSeason = useCallback((year: number, notes?: string) => {
     if (!data) return
 
-    if (USE_YJS_STORAGE) {
-      mutate(store => {
-        const now = new Date().toISOString()
-        const previousYear = year - 1
-        const previousSeason = store.seasons.find(s => s.year === previousYear)
+    mutate(store => {
+      const now = new Date().toISOString()
+      const previousYear = year - 1
+      const previousSeason = store.seasons.find(s => s.year === previousYear)
 
-        // Mirror the legacy area-season creation: one AreaSeason per
-        // non-archived area, rotating the rotation group for
-        // rotation-beds and copying grid positions where available.
-        const areaSeasons: AreaSeason[] = []
-        for (const area of store.areas) {
-          if (area.isArchived) continue
-          const previousAreaSeason = previousSeason?.areas.find(a => a.areaId === area.id)
-          let rotationGroup: RotationGroup | undefined
-          if (area.kind === 'rotation-bed') {
-            rotationGroup = previousAreaSeason?.rotationGroup
-              ? getNextRotationGroup(previousAreaSeason.rotationGroup)
-              : area.rotationGroup || 'legumes'
-          }
-          const gridPosition = previousAreaSeason?.gridPosition ?? area.gridPosition
-          areaSeasons.push(withoutUndefined({
-            areaId: area.id,
-            rotationGroup,
-            plantings: [],
-            gridPosition,
-          }))
+      // Mirror the legacy area-season creation: one AreaSeason per
+      // non-archived area, rotating the rotation group for
+      // rotation-beds and copying grid positions where available.
+      const areaSeasons: AreaSeason[] = []
+      for (const area of store.areas) {
+        if (area.isArchived) continue
+        const previousAreaSeason = previousSeason?.areas.find(a => a.areaId === area.id)
+        let rotationGroup: RotationGroup | undefined
+        if (area.kind === 'rotation-bed') {
+          rotationGroup = previousAreaSeason?.rotationGroup
+            ? getNextRotationGroup(previousAreaSeason.rotationGroup)
+            : area.rotationGroup || 'legumes'
         }
+        const gridPosition = previousAreaSeason?.gridPosition ?? area.gridPosition
+        areaSeasons.push(withoutUndefined({
+          areaId: area.id,
+          rotationGroup,
+          plantings: [],
+          gridPosition,
+        }))
+      }
 
-        const newSeason: SeasonRecord = withoutUndefined({
-          year,
-          status: 'planned',
-          areas: areaSeasons,
-          notes,
-          createdAt: now,
-          updatedAt: now,
-        })
-        store.seasons.push(newSeason)
-        store.state.currentYear = year
+      const newSeason: SeasonRecord = withoutUndefined({
+        year,
+        status: 'planned',
+        areas: areaSeasons,
+        notes,
+        createdAt: now,
+        updatedAt: now,
       })
-      setSelectedYear?.(year)
-      return
-    }
-
-    setData(addSeason(data, { year, status: 'planned', notes }))
+      store.seasons.push(newSeason)
+      store.state.currentYear = year
+    })
     setSelectedYear?.(year)
-  }, [data, setData, setSelectedYear, mutate])
+  }, [data, setSelectedYear, mutate])
 
   const deleteSeasonData = useCallback((year: number) => {
     if (!data) return
     // Don't allow deleting the last season
     if (data.seasons.length <= 1) return
 
-    if (USE_YJS_STORAGE) {
-      // Capture new currentYear synchronously so we can update the
-      // selected year after the transaction returns. The legacy path
-      // does this via the snapshot returned by removeSeason.
-      const yearNum = Number(year)
-      let newCurrentYear = data.currentYear
-      mutate(store => {
-        const idx = store.seasons.findIndex(s => Number(s.year) === yearNum)
-        if (idx === -1) return
-        store.seasons.splice(idx, 1)
+    // Capture new currentYear synchronously so we can update the
+    // selected year after the transaction returns.
+    const yearNum = Number(year)
+    let newCurrentYear = data.currentYear
+    mutate(store => {
+      const idx = store.seasons.findIndex(s => Number(s.year) === yearNum)
+      if (idx === -1) return
+      store.seasons.splice(idx, 1)
 
-        if (Number(store.state.currentYear) === yearNum) {
-          const years = store.seasons.map(s => Number(s.year)).sort((a, b) => b - a)
-          store.state.currentYear = years[0]
-          newCurrentYear = years[0]
-        }
-      })
-      setSelectedYear?.(newCurrentYear)
-      return
-    }
-
-    const newData = removeSeason(data, year)
-    setData(newData)
-    setSelectedYear?.(newData.currentYear)
-  }, [data, setData, setSelectedYear, mutate])
+      if (Number(store.state.currentYear) === yearNum) {
+        const years = store.seasons.map(s => Number(s.year)).sort((a, b) => b - a)
+        store.state.currentYear = years[0]
+        newCurrentYear = years[0]
+      }
+    })
+    setSelectedYear?.(newCurrentYear)
+  }, [data, setSelectedYear, mutate])
 
   const updateSeasonNotes = useCallback((notes: string) => {
     if (!data) return
 
-    if (USE_YJS_STORAGE) {
-      mutate(store => {
-        const season = store.seasons.find(s => s.year === selectedYear)
-        if (!season) return
-        season.notes = notes
-        season.updatedAt = new Date().toISOString()
-      })
-      return
-    }
-
-    setData(updateSeason(data, selectedYear, { notes }))
-  }, [data, selectedYear, setData, mutate])
+    mutate(store => {
+      const season = store.seasons.find(s => s.year === selectedYear)
+      if (!season) return
+      season.notes = notes
+      season.updatedAt = new Date().toISOString()
+    })
+  }, [data, selectedYear, mutate])
 
   // ============ LAYOUT HELPERS ============
